Move Constants.cs files from single Constants folder to appropriate rule folder.
Update Namespace of moved Constants.cs files.
