+++ src/nunit.analyzers/ContainsConstraintWrongActualType/ContainsConstraintWrongActualTypeConstants.cs
@@ -1,4 +1,4 @@
-namespace NUnit.Analyzers.Constants
+namespace NUnit.Analyzers.ContainsConstraintWrongActualType
 {
     internal static class ContainsConstraintWrongActualTypeConstants
     {
