+++ src/nunit.analyzers/IgnoreCaseUsage/IgnoreCaseUsageAnalyzerConstants.cs
@@ -1,4 +1,4 @@
-namespace NUnit.Analyzers.Constants
+namespace NUnit.Analyzers.IgnoreCaseUsage
 {
     internal static class IgnoreCaseUsageAnalyzerConstants
     {
